more test fixing and stuff
