+++ conftest.py
@@ -1,0 +1,3 @@
+import matplotlib
+
+matplotlib.use("Agg")
